Updates to build process
- Includes sample buildpack.toml files that are modified to work with ARM64
- Removes builder.toml and replaces with a script that will generate one
- Various small fixes to the scripts
- Update Dockerfile to take the stack id as an argument

+++ scripts/clone.sh
@@ -3,7 +3,7 @@ set -eo pipefail
 
 BPID="$1"
 BPVER="$2"
-WORK="./buildpacks-tmp"
+WORK="./buildpacks"
 
 if [ -z "$BPID" ] || [ -z "$BPVER" ]; then
 	echo ""
@@ -22,7 +22,7 @@ if [ -z "$WORK" ]; then
 fi
 
 mkdir -p "$WORK"
-rm -rf "$WORK/"*
+rm -rf "${WORK:?}/"*
 
 git clone "https://github.com/$BPID" "$WORK/$BPID"
 pushd "$WORK/$BPID" >/dev/null
